remove old borked version from the index
Its version number (0.6) is higher than the more recent helm charts, because it was a mistake.
Therefore, when installing alumet from the repo, it will automatically pull this old version by default, which is not what we want.
Exceptionnaly remove this version (manually) to restore the well-functioning of the repo.